requirements.txt
asgiref==3.3.4
attrs==21.2.0
click==8.0.1
colorama==0.4.4
Django==3.2.4
dlib==19.17.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.1
flask-swagger-ui==3.36.0
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
mistune==0.8.4
mysql-connector-python==8.0.25
numpy==1.19.5
opencv-python==4.5.1.48
pandas==1.1.5
Pillow==8.2.0
protobuf==3.17.2
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
scikit==0.24.0
scikit-build==0.11.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1